updates on work with EEPROM memory

+++ proyecto-final/test/eeprom.h
@@ -4,7 +4,7 @@
 #define EEPROM
 
 void memory_init();
-void memory_write(uint8_t address, uint8_t data);
-uint8_t memory_read(uint8_t address);
+void memory_write(uint16_t address, uint8_t data);
+uint8_t memory_read(uint16_t address);
 
 #endif
